modularized build.rs in buildrs directory
