+++ buildrs/cbindgen.rs
@@ -1,36 +1,54 @@
-use std::borrow::Cow;
-use std::collections::HashSet;
-use std::path::Path;
-use std::{collections::HashMap, fs::File};
-use std::io::{BufRead, BufWriter, Read, Write};
+use std::{collections::HashSet, fs::File, io::{BufRead, Read, Write}, path::{Path, PathBuf}};
+
 use fs2::FileExt;
+use phf::phf_map;
 use regex::Regex;
 
-use super::{cargo_target_dir, split_type_name, test_feature, RUST_TO_C_FEATURES};
+use super::{cargo_target_dir, split_bindings, split_type_name, test_feature, FuncArg, FunctionSignature};
 
 const BUGGY_GENERATION_PATH: &str = "include/zenoh-gen-buggy.h";
 const GENERATION_PATH: &str = "include/zenoh-gen.h";
 const PREPROCESS_PATH: &str = "include/zenoh-cpp.h";
-const SPLITGUIDE_PATH: &str = "splitguide.yaml";
-const HEADER: &str = r"//
-// Copyright (c) 2024 ZettaScale Technology
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License 2.0 which is available at
-// http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-// which is available at https://www.apache.org/licenses/LICENSE-2.0.
-//
-// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-//
-// Contributors:
-//   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
-//
-// clang-format off
-#ifdef DOCS
-#define ALIGN(n)
-#define ZENOHC_API
-#endif
-";
+
+static RUST_TO_C_FEATURES: phf::Map<&'static str, &'static str> = phf_map! {
+    "unstable" => "Z_FEATURE_UNSTABLE_API",
+    "shared-memory" => "Z_FEATURE_SHARED_MEMORY",
+    "auth_pubkey" => "Z_FEATURE_AUTH_PUBKEY",
+    "auth_usrpwd" => "Z_FEATURE_AUTH_USRPWD",
+    "transport_multilink" => "Z_FEATURE_TRANSPORT_MULTILINK",
+    "transport_compression" => "Z_FEATURE_TRANSPORT_COMPRESSION",
+    "transport_quic"  => "Z_FEATURE_TRANSPORT_QUIC",
+    "transport_tcp" =>  "Z_FEATURE_TRANSPORT_TCP",
+    "transport_tls" =>  "Z_FEATURE_TRANSPORT_TLS",
+    "transport_udp" =>  "Z_FEATURE_TRANSPORT_UDP",
+    "transport_unixsock-stream" =>  "Z_FEATURE_TRANSPORT_UNIXSOCK_STREAM",
+    "transport_ws" =>  "Z_FEATURE_TRANSPORT_WS",
+    "transport_vsock" => "Z_FEATURE_VSOCK"
+};
+
+pub fn generate_c_headers() {
+    cbindgen::generate(std::env::var("CARGO_MANIFEST_DIR").unwrap())
+        .expect("Unable to generate bindings")
+        .write_to_file(BUGGY_GENERATION_PATH);
+
+    fix_cbindgen(BUGGY_GENERATION_PATH, GENERATION_PATH);
+    std::fs::remove_file(BUGGY_GENERATION_PATH).unwrap();
+
+    preprocess_header(GENERATION_PATH, PREPROCESS_PATH);
+    create_generics_header(PREPROCESS_PATH, "include/zenoh_macros.h");
+    std::fs::remove_file(PREPROCESS_PATH).unwrap();
+
+    configure();
+    let files= split_bindings(GENERATION_PATH).unwrap();
+    text_replace(files.iter());
+
+    fs_extra::copy_items(
+        &["include"],
+        cargo_target_dir(),
+        &fs_extra::dir::CopyOptions::default().overwrite(true),
+    )
+    .expect("include should be copied to CARGO_TARGET_DIR");
+}
 
 fn fix_cbindgen(input: &str, output: &str) {
     let bindings = std::fs::read_to_string(input).expect("failed to open input file");
@@ -46,724 +64,7 @@ fn preprocess_header(input: &str, output: &str) {
     out.write_all(parsed.as_bytes()).unwrap();
 }
 
-fn configure() {
-    let mut file = std::fs::File::options()
-        .write(true)
-        .truncate(true)
-        .append(false)
-        .create(true)
-        .open("include/zenoh_configure.h")
-        .unwrap();
-    file.lock_exclusive().unwrap();
-
-    let version = std::fs::read_to_string("version.txt").unwrap();
-    let version = version.trim();
-    let version_parts: Vec<&str> = version.split('.').collect();
-    if version_parts.len() < 3 {
-        panic!("Invalid version format: \"{}\" in file version.txt. Major.Minor.Patch parts are required", version);
-    }
-    let major = version_parts[0];
-    let minor = version_parts[1];
-    let patch = version_parts[2];
-    let tweak = version_parts.get(3).unwrap_or(&"");
-    file.write_all(
-        format!(
-            r#"#pragma once
-#define ZENOH_C "{}"
-#define ZENOH_C_MAJOR {}
-#define ZENOH_C_MINOR {}
-#define ZENOH_C_PATCH {}
-#define ZENOH_C_TWEAK {}
-
-#define TARGET_ARCH_{}
-"#,
-            version,
-            major,
-            minor,
-            patch,
-            tweak,
-            std::env::var("CARGO_CFG_TARGET_ARCH")
-                .unwrap()
-                .to_uppercase()
-        )
-        .as_bytes(),
-    )
-    .unwrap();
-
-    for (rust_feature, c_feature) in RUST_TO_C_FEATURES.entries() {
-        if test_feature(rust_feature) {
-            file.write_all(format!("#define {}\n", c_feature).as_bytes())
-                .unwrap();
-        }
-    }
-    fs2::FileExt::unlock(&file).unwrap();
-}
-
-fn text_replace<'a>(files: impl Iterator<Item = &'a str>) {
-    for name in files {
-        let path = format!("include/{}", name);
-
-        // Read content
-        let mut file = std::fs::File::options()
-            .read(true)
-            .create(false)
-            .write(false)
-            .open(&path)
-            .unwrap();
-        file.lock_exclusive().unwrap();
-        let mut buf = String::new();
-        file.read_to_string(&mut buf).unwrap();
-        fs2::FileExt::unlock(&file).unwrap();
-
-        // Remove _T_ from enum variant name
-        let buf = buf.replace("_T_", "_");
-        // Replace _t_Tag from union variant name
-        let buf = buf.replace("_t_Tag", "_tag_t");
-        // Insert `ZENOHC_API` macro before `extern const`.
-        // The cbindgen can prefix functions (see `[fn] prefix=...` in cbindgn.toml), but not extern variables.
-        // So have to do it here.
-        let buf = buf.replace("extern const", "ZENOHC_API extern const");
-
-        // Overwrite content
-        let mut file = std::fs::File::options()
-            .read(false)
-            .create(false)
-            .write(true)
-            .truncate(true)
-            .open(&path)
-            .unwrap();
-        file.lock_exclusive().unwrap();
-        file.write_all(buf.as_bytes()).unwrap();
-        fs2::FileExt::unlock(&file).unwrap();
-    }
-}
-
-fn split_bindings(split_guide: &SplitGuide) -> Result<(), String> {
-    let bindings = std::fs::read_to_string(GENERATION_PATH).unwrap();
-    let mut files = split_guide
-        .rules
-        .iter()
-        .map(|(name, _)| {
-            let file = std::fs::File::options()
-                .write(true)
-                .truncate(true)
-                .append(false)
-                .create(true)
-                .open(format!("include/{}", name))
-                .unwrap();
-            file.lock_exclusive().unwrap();
-            file.set_len(0).unwrap();
-            (name.as_str(), BufWriter::new(file))
-        })
-        .collect::<HashMap<_, _>>();
-    for file in files.values_mut() {
-        file.write_all(HEADER.as_bytes())
-            .map_err(|e| e.to_string())?;
-    }
-    let mut records = group_tokens(Tokenizer {
-        filename: GENERATION_PATH,
-        inner: &bindings,
-    })?;
-    for id in split_guide.requested_ids() {
-        if !records.iter().any(|r| r.contains_id(id)) {
-            return Err(format!(
-                "{} not found (requested explicitly by splitguide.yaml)",
-                id,
-            ));
-        }
-    }
-    for record in &mut records {
-        let appropriate_files = split_guide.appropriate_files(record);
-        for file in appropriate_files {
-            let writer = files.get_mut(file).unwrap();
-            record.used = true;
-            write!(writer, "{}", &record).unwrap();
-        }
-    }
-    for record in &records {
-        record.is_used()?;
-    }
-    for (_, file) in files {
-        fs2::FileExt::unlock(&file.into_inner().unwrap()).unwrap();
-    }
-    std::fs::remove_file(GENERATION_PATH).unwrap();
-    Ok(())
-}
-
-enum SplitRule {
-    Brand(RecordType),
-    Exclusive(String),
-    Shared(String),
-}
-struct SplitGuide {
-    rules: Vec<(String, Vec<SplitRule>)>,
-}
-impl SplitGuide {
-    fn from_yaml<P: AsRef<Path>>(path: P) -> Self {
-        let map: HashMap<String, Vec<String>> =
-            serde_yaml::from_reader(std::fs::File::open(path).unwrap()).unwrap();
-        SplitGuide {
-            rules: map
-                .into_iter()
-                .map(|(name, rules)| {
-                    (
-                        name,
-                        rules
-                            .into_iter()
-                            .filter_map(|s| {
-                                let mut split = s.split('#');
-                                let val = split.next().unwrap();
-                                for feature in split {
-                                    if !test_feature(feature) {
-                                        return None;
-                                    }
-                                }
-                                Some(val.to_owned())
-                            })
-                            .map(|mut s| match s.as_str() {
-                                ":functions" => SplitRule::Brand(RecordType::Function),
-                                ":typedefs" => SplitRule::Brand(RecordType::Typedef),
-                                ":includes" => SplitRule::Brand(RecordType::PreprInclude),
-                                ":defines" => SplitRule::Brand(RecordType::PreprDefine),
-                                ":const" => SplitRule::Brand(RecordType::Const),
-                                ":multiples" => SplitRule::Brand(RecordType::Multiple),
-                                _ if s.ends_with('!') => {
-                                    s.pop();
-                                    SplitRule::Exclusive(s)
-                                }
-                                _ => SplitRule::Shared(s),
-                            })
-                            .collect(),
-                    )
-                })
-                .collect(),
-        }
-    }
-    fn appropriate_files(&self, record: &Record) -> Vec<&str> {
-        let mut shared = Vec::new();
-        let mut exclusives = Vec::new();
-        for (file, rules) in &self.rules {
-            for rule in rules {
-                match rule {
-                    SplitRule::Brand(brand) if *brand == record.rt => shared.push(file.as_str()),
-                    SplitRule::Exclusive(id) if record.contains_id(id) => {
-                        exclusives.push(file.as_str())
-                    }
-                    SplitRule::Shared(id) if record.contains_id(id) => shared.push(file.as_str()),
-                    _ => {}
-                }
-            }
-        }
-        if exclusives.is_empty() {
-            shared
-        } else {
-            exclusives
-        }
-    }
-    fn requested_ids(&self) -> impl Iterator<Item = &str> {
-        self.rules.iter().flat_map(|(_, rules)| {
-            rules.iter().filter_map(|r| match r {
-                SplitRule::Brand(_) => None,
-                SplitRule::Exclusive(s) | SplitRule::Shared(s) => Some(s.as_str()),
-            })
-        })
-    }
-}
-
-fn group_tokens(stream: Tokenizer) -> Result<Vec<Record>, String> {
-    let mut records = Vec::new();
-    let mut record_collect = Record::new();
-    for token in stream {
-        record_collect.add_token(token)?;
-        if record_collect.is_ready() {
-            let mut record = Record::new();
-            std::mem::swap(&mut record_collect, &mut record);
-            records.push(record);
-        }
-    }
-    records.push(record_collect);
-    Ok(records)
-}
-
-#[derive(Copy, Clone, PartialEq, Debug)]
-enum RecordType {
-    Empty,
-    Multiple,
-    PrivateToken,
-    Typedef,
-    Function,
-    Const,
-    PreprDefine,
-    PreprInclude,
-}
-
-impl RecordType {
-    fn update(&mut self, rt: RecordType) {
-        match *self {
-            RecordType::Empty => *self = rt,
-            RecordType::Multiple => {}
-            _ => *self = RecordType::Multiple,
-        }
-    }
-}
-
-struct Record<'a> {
-    used: bool,
-    rt: RecordType,
-    nesting: i32,
-    ids: Vec<Cow<'a, str>>,
-    tokens: Vec<Token<'a>>,
-}
-
-impl<'a> Record<'a> {
-    fn new() -> Self {
-        Self {
-            used: false,
-            rt: RecordType::Empty,
-            nesting: 0,
-            ids: Vec::new(),
-            tokens: Vec::new(),
-        }
-    }
-
-    fn is_used(&self) -> Result<(), String> {
-        if self.used || self.rt == RecordType::Empty || self.rt == RecordType::PrivateToken {
-            Ok(())
-        } else {
-            let token_ids = self.tokens.iter().map(|t| t.id).collect::<Vec<_>>();
-            Err(format!("Unused {:?} record: {:?}", self.rt, token_ids))
-        }
-    }
-
-    fn is_ready(&self) -> bool {
-        self.nesting == 0 && self.rt != RecordType::Empty
-    }
-
-    fn contains_id(&self, id: &str) -> bool {
-        self.ids.iter().any(|v| v == id)
-    }
-
-    fn push_token(&mut self, token: Token<'a>) {
-        self.tokens.push(token);
-    }
-
-    fn push_record_type_token(&mut self, token: Token<'a>, rt: RecordType) {
-        self.rt.update(rt);
-        if !token.id.is_empty() {
-            self.ids.push(token.id.into());
-        }
-        self.push_token(token)
-    }
-
-    fn push_prepr_if(&mut self, token: Token<'a>) {
-        self.nesting += 1;
-        self.push_token(token)
-    }
-
-    fn push_prepr_endif(&mut self, token: Token<'a>) -> Result<(), String> {
-        self.nesting -= 1;
-        if self.nesting < 0 {
-            return Err("unmatched #endif".into());
-        }
-        self.push_token(token);
-        Ok(())
-    }
-
-    fn add_token(&mut self, token: Token<'a>) -> Result<(), String> {
-        match token.tt {
-            TokenType::Comment => self.push_token(token),
-            TokenType::Typedef => self.push_record_type_token(token, RecordType::Typedef),
-            TokenType::Function => self.push_record_type_token(token, RecordType::Function),
-            TokenType::Const => self.push_record_type_token(token, RecordType::Const),
-            TokenType::PrivateToken => self.push_record_type_token(token, RecordType::PrivateToken),
-            TokenType::PreprDefine => self.push_record_type_token(token, RecordType::PreprDefine),
-            TokenType::PreprInclude => self.push_record_type_token(token, RecordType::PreprInclude),
-            TokenType::PreprIf => self.push_prepr_if(token),
-            TokenType::PreprElse => self.push_token(token),
-            TokenType::PreprEndif => self.push_prepr_endif(token)?,
-            TokenType::Whitespace => self.push_token(token),
-        }
-        Ok(())
-    }
-}
-
-// Print all comments first, skip whitespaces
-impl std::fmt::Display for Record<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.tokens
-            .iter()
-            .filter(|t| t.tt == TokenType::Comment)
-            .map(|t| t.fmt(f))
-            .find(|r| r.is_err())
-            .unwrap_or(Ok(()))?;
-        self.tokens
-            .iter()
-            .filter(|t| t.tt != TokenType::Comment && t.tt != TokenType::Whitespace)
-            .map(|t| t.fmt(f))
-            .find(|r| r.is_err())
-            .unwrap_or(Ok(()))?;
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum TokenType {
-    Comment,
-    Typedef,
-    Function,
-    Const,
-    PrivateToken,
-    PreprDefine,
-    PreprInclude,
-    PreprIf,
-    PreprElse,
-    PreprEndif,
-    Whitespace,
-}
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct Token<'a> {
-    tt: TokenType,
-    id: &'a str,
-    span: Cow<'a, str>,
-}
-impl std::fmt::Display for Token<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // f.write_str(format!("{:?} [", self.tt).as_str())?;
-        f.write_str(&self.span)?;
-        // f.write_str("]")?;
-
-        // Each token is finalized with endline character on output
-        f.write_str("\n")?;
-        Ok(())
-    }
-}
-impl<'a> Token<'a> {
-    fn new<I: Into<Cow<'a, str>>>(tt: TokenType, id: &'a str, span: I) -> Self {
-        Token {
-            tt,
-            id,
-            span: span.into(),
-        }
-    }
-
-    fn next(s: &'a str) -> Option<Self> {
-        Self::whitespace(s)
-            .or_else(|| Self::comment(s))
-            .or_else(|| Self::prepr_endif(s))
-            .or_else(|| Self::prepr_include(s))
-            .or_else(|| Self::prepr_define(s))
-            .or_else(|| Self::prepr_if(s))
-            .or_else(|| Self::prepr_else(s))
-            .or_else(|| Self::typedef(s))
-            .or_else(|| Self::r#const(s))
-            .or_else(|| Self::function(s))
-    }
-
-    //
-    // Each token is consumed without end of line characters.
-    // When performing output of tokens, endline is added to each token.
-    // This guarantees that tokens can be shuffled as necessary
-    //
-    fn typedef(s: &'a str) -> Option<Self> {
-        if s.starts_with("typedef") {
-            let mut len = 0;
-            let mut accolades = 0;
-            for c in s.chars() {
-                len += c.len_utf8();
-                if c == '{' {
-                    accolades += 1;
-                } else if c == '}' {
-                    accolades -= 1;
-                } else if c == ';' && accolades == 0 {
-                    break;
-                }
-            }
-            let span = &s[..len];
-            let id_len = span
-                .chars()
-                .rev()
-                .skip(1)
-                .take_while(|&c| c.is_alphanumeric() || c == '_')
-                .fold(1, |acc, c| acc + c.len_utf8());
-            let id = &span[(span.len() - id_len)..(span.len() - 1)];
-            Some(Token::new(
-                if id.starts_with('_') {
-                    TokenType::PrivateToken
-                } else {
-                    TokenType::Typedef
-                },
-                id,
-                span,
-            ))
-        } else {
-            None
-        }
-    }
-
-    fn function(s: &'a str) -> Option<Self> {
-        s.until_incl(";").and_then(|span| {
-            span.contains('(').then(|| {
-                let mut iter = span.chars().rev();
-                let id_end = iter
-                    .by_ref()
-                    .take_while(|&c| c != '(')
-                    .fold(1, |acc, c| acc + c.len_utf8());
-                let id_len = iter
-                    .take_while(|&c| c.is_alphanumeric() || c == '_')
-                    .fold(0, |acc, c| acc + c.len_utf8());
-                let id = &span[(span.len() - id_end - id_len)..(span.len() - id_end)];
-                Token::new(TokenType::Function, id, span)
-            })
-        })
-    }
-
-    fn r#const(s: &'a str) -> Option<Self> {
-        s.until_incl(";").and_then(|span| {
-            span.contains("const").then(|| {
-                let id_len = span
-                    .chars()
-                    .rev()
-                    .skip(1)
-                    .take_while(|&c| c.is_alphanumeric() || c == '_')
-                    .fold(0, |acc, c| acc + c.len_utf8());
-                let id = &span[(span.len() - 1 - id_len)..(span.len() - 1)];
-                Token::new(TokenType::Const, id, span)
-            })
-        })
-    }
-
-    fn whitespace(s: &'a str) -> Option<Self> {
-        let mut len = 0;
-        for c in s.chars() {
-            if c.is_whitespace() {
-                len += c.len_utf8()
-            } else {
-                break;
-            }
-        }
-        if len > 0 {
-            Some(Token::new(TokenType::Whitespace, "", &s[..len]))
-        } else {
-            None
-        }
-    }
-
-    fn comment(s: &'a str) -> Option<Self> {
-        if s.starts_with("/*") {
-            Some(Token::new(
-                TokenType::Comment,
-                "",
-                s.until_incl("*/").unwrap_or(s),
-            ))
-        } else if s.starts_with("//") {
-            Some(Token::new(
-                TokenType::Comment,
-                "",
-                s.until("\n").unwrap_or(s),
-            ))
-        } else {
-            None
-        }
-    }
-
-    fn prepr_if(s: &'a str) -> Option<Self> {
-        if s.starts_with("#if ") || s.starts_with("#ifdef ") || s.starts_with("#ifndef ") {
-            let span = s.until("\n").unwrap_or(s);
-            Some(Token::new(TokenType::PreprIf, span, span))
-        } else {
-            None
-        }
-    }
-
-    fn prepr_define(s: &'a str) -> Option<Self> {
-        let start = "#define ";
-        s.strip_prefix(start).map(|defined| {
-            let span = s.until("\n").unwrap_or(s);
-            Token::new(
-                if defined.starts_with('_') {
-                    TokenType::PrivateToken
-                } else {
-                    TokenType::PreprDefine
-                },
-                span[start.len()..].split_whitespace().next().unwrap(),
-                span,
-            )
-        })
-    }
-
-    fn prepr_endif(s: &'a str) -> Option<Self> {
-        s.starts_with("#endif")
-            .then(|| Token::new(TokenType::PreprEndif, "", s.until("\n").unwrap_or(s)))
-    }
-
-    fn prepr_else(s: &'a str) -> Option<Self> {
-        s.starts_with("#else")
-            .then(|| Token::new(TokenType::PreprElse, "", s.until("\n").unwrap_or(s)))
-    }
-
-    fn prepr_include(s: &'a str) -> Option<Self> {
-        Self::r#include(s, "#include \"", "\"").or_else(|| Self::r#include(s, "#include <", ">"))
-    }
-
-    fn r#include(s: &'a str, start: &str, end: &str) -> Option<Self> {
-        if s.starts_with(start) {
-            let span = s.until_incl(end).expect("detected unterminated #include");
-            Some(Token::new(
-                TokenType::PreprInclude,
-                &span[start.len()..(span.len() - end.len())],
-                span,
-            ))
-        } else {
-            None
-        }
-    }
-}
-trait Until: Sized {
-    fn until(self, pattern: &str) -> Option<Self>;
-    fn until_incl(self, pattern: &str) -> Option<Self>;
-}
-impl Until for &str {
-    fn until(self, pattern: &str) -> Option<Self> {
-        self.find(pattern).map(|l| &self[..l])
-    }
-    fn until_incl(self, pattern: &str) -> Option<Self> {
-        self.find(pattern).map(|l| &self[..(l + pattern.len())])
-    }
-}
-struct Tokenizer<'a> {
-    filename: &'a str,
-    inner: &'a str,
-}
-impl<'a> Iterator for Tokenizer<'a> {
-    type Item = Token<'a>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.inner.is_empty() {
-            None
-        } else {
-            let result = Token::next(self.inner);
-            if let Some(result) = &result {
-                self.inner = &self.inner[result.span.len()..];
-            } else {
-                panic!(
-                    "Couldn't parse C file {}, stopped at: {}",
-                    self.filename,
-                    self.inner.lines().next().unwrap()
-                )
-            }
-            result
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct Ctype {
-    typename: String,
-}
-
-impl Ctype {
-    pub fn new(typename: &str) -> Self {
-        Ctype {
-            typename: typename.to_owned(),
-        }
-    }
-
-    pub fn without_cv(self) -> Self {
-        Ctype {
-            typename: self.typename.replace("const ", ""),
-        }
-    }
-
-    pub fn without_ptr(self) -> Self {
-        Ctype {
-            typename: self.typename.replace(['*', '&'], ""),
-        }
-    }
-
-    pub fn with_ref(self) -> Self {
-        Ctype {
-            typename: self.typename.replace('*', "&"),
-        }
-    }
-
-    pub fn decay(self) -> Self {
-        self.without_cv().without_ptr()
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct FuncArg {
-    typename: Ctype,
-    name: String,
-}
-
-impl FuncArg {
-    pub fn new(typename: &str, name: &str) -> Self {
-        FuncArg {
-            typename: Ctype::new(typename),
-            name: name.to_owned(),
-        }
-    }
-}
-#[derive(Clone, Debug)]
-pub struct FunctionSignature {
-    entity_name: String, // the signifcant part of name, e.g. `session` for `z_session_t`
-    return_type: Ctype,
-    func_name: String,
-    args: Vec<FuncArg>,
-}
-
-impl FunctionSignature {
-    pub fn new(
-        entity_name: &str,
-        return_type: &str,
-        func_name: String,
-        args: Vec<FuncArg>,
-    ) -> Self {
-        FunctionSignature {
-            entity_name: entity_name.to_owned(),
-            return_type: Ctype::new(return_type),
-            func_name,
-            args,
-        }
-    }
-}
-
-
-
-
-/// Evaluates conditional feature macros in the form #if (logical expression of define(FEATURE_NAME))
-/// and removes the code under those that evaluate to false
-/// Note: works only on single string conditional expressions
-pub fn process_feature_defines(input_path: &str) -> Result<String, Box<dyn std::error::Error>> {
-    let file = std::fs::File::open(input_path)?;
-    let lines = std::io::BufReader::new(file).lines();
-    let mut out = String::new();
-    let mut skip = false;
-    let mut nest_level: usize = 0;
-    for line in lines.map_while(Result::ok) {
-        if line.starts_with("#ifdef") && skip {
-            nest_level += 1;
-        } else if line.starts_with("#endif") && skip {
-            nest_level -= 1;
-            skip = nest_level != 0;
-            continue;
-        } else if line.starts_with("#if ") {
-            skip = skip || evaluate_c_defines_line(&line);
-            if skip {
-                nest_level += 1;
-            }
-        }
-        if !skip {
-            out += &line;
-            out += "\n";
-        }
-    }
-
-    Ok(out)
-}
-
-pub fn create_generics_header(path_in: &str, path_out: &str) {
+fn create_generics_header(path_in: &str, path_out: &str) {
     let mut file_out = std::fs::File::options()
         .read(false)
         .write(true)
@@ -997,8 +298,149 @@ pub fn create_generics_header(path_in: &str, path_out: &str) {
         .unwrap();
 }
 
+fn configure() {
+    let mut file = std::fs::File::options()
+        .write(true)
+        .truncate(true)
+        .append(false)
+        .create(true)
+        .open("include/zenoh_configure.h")
+        .unwrap();
+    file.lock_exclusive().unwrap();
 
-pub fn make_move_take_signatures(
+    let version = std::fs::read_to_string("version.txt").unwrap();
+    let version = version.trim();
+    let version_parts: Vec<&str> = version.split('.').collect();
+    if version_parts.len() < 3 {
+        panic!("Invalid version format: \"{}\" in file version.txt. Major.Minor.Patch parts are required", version);
+    }
+    let major = version_parts[0];
+    let minor = version_parts[1];
+    let patch = version_parts[2];
+    let tweak = version_parts.get(3).unwrap_or(&"");
+    file.write_all(
+        format!(
+            r#"#pragma once
+#define ZENOH_C "{}"
+#define ZENOH_C_MAJOR {}
+#define ZENOH_C_MINOR {}
+#define ZENOH_C_PATCH {}
+#define ZENOH_C_TWEAK {}
+
+#define TARGET_ARCH_{}
+"#,
+            version,
+            major,
+            minor,
+            patch,
+            tweak,
+            std::env::var("CARGO_CFG_TARGET_ARCH")
+                .unwrap()
+                .to_uppercase()
+        )
+        .as_bytes(),
+    )
+    .unwrap();
+
+    for (rust_feature, c_feature) in RUST_TO_C_FEATURES.entries() {
+        if test_feature(rust_feature) {
+            file.write_all(format!("#define {}\n", c_feature).as_bytes())
+                .unwrap();
+        }
+    }
+    fs2::FileExt::unlock(&file).unwrap();
+}
+
+fn text_replace(files: impl Iterator<Item = impl AsRef<Path>>) {
+    for name in files {
+        let path = PathBuf::from("include").join(name);
+
+        // Read content
+        let mut file = std::fs::File::options()
+            .read(true)
+            .create(false)
+            .write(false)
+            .open(&path)
+            .unwrap();
+        file.lock_exclusive().unwrap();
+        let mut buf = String::new();
+        file.read_to_string(&mut buf).unwrap();
+        fs2::FileExt::unlock(&file).unwrap();
+
+        // Remove _T_ from enum variant name
+        let buf = buf.replace("_T_", "_");
+        // Replace _t_Tag from union variant name
+        let buf = buf.replace("_t_Tag", "_tag_t");
+        // Insert `ZENOHC_API` macro before `extern const`.
+        // The cbindgen can prefix functions (see `[fn] prefix=...` in cbindgn.toml), but not extern variables.
+        // So have to do it here.
+        let buf = buf.replace("extern const", "ZENOHC_API extern const");
+
+        // Overwrite content
+        let mut file = std::fs::File::options()
+            .read(false)
+            .create(false)
+            .write(true)
+            .truncate(true)
+            .open(&path)
+            .unwrap();
+        file.lock_exclusive().unwrap();
+        file.write_all(buf.as_bytes()).unwrap();
+        fs2::FileExt::unlock(&file).unwrap();
+    }
+}
+
+/// Evaluates conditional feature macros in the form #if (logical expression of define(FEATURE_NAME))
+/// and removes the code under those that evaluate to false
+/// Note: works only on single string conditional expressions
+fn process_feature_defines(input_path: &str) -> Result<String, Box<dyn std::error::Error>> {
+    let file = std::fs::File::open(input_path)?;
+    let lines = std::io::BufReader::new(file).lines();
+    let mut out = String::new();
+    let mut skip = false;
+    let mut nest_level: usize = 0;
+    for line in lines.map_while(Result::ok) {
+        if line.starts_with("#ifdef") && skip {
+            nest_level += 1;
+        } else if line.starts_with("#endif") && skip {
+            nest_level -= 1;
+            skip = nest_level != 0;
+            continue;
+        } else if line.starts_with("#if ") {
+            skip = skip || evaluate_c_defines_line(&line);
+            if skip {
+                nest_level += 1;
+            }
+        }
+        if !skip {
+            out += &line;
+            out += "\n";
+        }
+    }
+
+    Ok(out)
+}
+
+fn evaluate_c_defines_line(line: &str) -> bool {
+    let mut s = line.to_string();
+    for (rust_feature, c_feature) in RUST_TO_C_FEATURES.entries() {
+        s = s.replace(
+            &format!("defined({})", c_feature),
+            match test_feature(rust_feature) {
+                true => "true",
+                false => "false",
+            },
+        );
+    }
+
+    s = s.replace("#if", "");
+    match evalexpr::eval(&s) {
+        Ok(v) => v == evalexpr::Value::from(false),
+        Err(_) => panic!("Failed to evaluate {}", &s),
+    }
+}
+
+fn make_move_take_signatures(
     path_in: &str,
 ) -> (Vec<FunctionSignature>, Vec<FunctionSignature>) {
     let bindings = std::fs::read_to_string(path_in).unwrap();
@@ -1033,7 +475,7 @@ pub fn make_move_take_signatures(
     (move_funcs, take_funcs)
 }
 
-pub fn find_loan_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_loan_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(r"const struct (\w+) \*(\w+)_loan\(const struct (\w+) \*(\w+)\);").unwrap();
     let mut res = Vec::<FunctionSignature>::new();
@@ -1056,7 +498,7 @@ pub fn find_loan_functions(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn find_loan_mut_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_loan_mut_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(r"struct (\w+) \*(\w+)_loan_mut\(struct (\w+) \*(\w+)\);").unwrap();
     let mut res = Vec::<FunctionSignature>::new();
@@ -1076,7 +518,7 @@ pub fn find_loan_mut_functions(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn find_take_from_loaned_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_take_from_loaned_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(r"void (\w+)_take_from_loaned\(struct (\w+) \*(\w+)").unwrap();
     let mut res = Vec::<FunctionSignature>::new();
@@ -1099,7 +541,7 @@ pub fn find_take_from_loaned_functions(path_in: &str) -> Vec<FunctionSignature> 
     res
 }
 
-pub fn find_drop_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_drop_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(r"(.+?) +(\w+_drop)\(struct (\w+) \*(\w+)\);").unwrap();
     let mut res = Vec::<FunctionSignature>::new();
@@ -1126,7 +568,7 @@ pub fn find_drop_functions(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn find_null_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_null_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(r" (z.?_internal_\w+_null)\(struct (\w+) \*(\w+)\);").unwrap();
     let mut res = Vec::<FunctionSignature>::new();
@@ -1144,7 +586,7 @@ pub fn find_null_functions(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn find_check_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_check_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(r"bool (z.?_internal_\w+_check)\(const struct (\w+) \*(\w+)\);").unwrap();
     let mut res = Vec::<FunctionSignature>::new();
@@ -1165,7 +607,7 @@ pub fn find_check_functions(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn find_call_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_call_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(
         r"(\w+) (\w+)_call\(const struct (\w+) \*(\w+),\s+(\w*)\s*struct (\w+) (\*?)(\w+)\);",
@@ -1198,7 +640,7 @@ pub fn find_call_functions(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn find_closure_constructors(path_in: &str) -> Vec<FunctionSignature> {
+fn find_closure_constructors(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(
         r"(\w+) (\w+)_closure_(\w+)\(struct\s+(\w+)\s+\*(\w+),\s+void\s+\(\*call\)(\([\s\w,\*]*\)),\s+void\s+\(\*drop\)(\(.*\)),\s+void\s+\*context\);"
@@ -1234,7 +676,7 @@ pub fn find_closure_constructors(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn find_recv_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_recv_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(r"(\w+)\s+z_(\w+)_handler_(\w+)_recv\(const\s+struct\s+(\w+)\s+\*(\w+),\s+struct\s+(\w+)\s+\*(\w+)\);").unwrap();
     let mut res = Vec::<FunctionSignature>::new();
@@ -1257,7 +699,7 @@ pub fn find_recv_functions(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn find_clone_functions(path_in: &str) -> Vec<FunctionSignature> {
+fn find_clone_functions(path_in: &str) -> Vec<FunctionSignature> {
     let bindings = std::fs::read_to_string(path_in).unwrap();
     let re = Regex::new(
         r"(\w+)\s+z_(\w+)_clone\(struct\s+(\w+)\s+\*(\w+),\s+const\s+struct\s+(\w+)\s+\*(\w+)\);",
@@ -1283,7 +725,7 @@ pub fn find_clone_functions(path_in: &str) -> Vec<FunctionSignature> {
     res
 }
 
-pub fn generate_generic_c(
+fn generate_generic_c(
     macro_func: &[FunctionSignature],
     generic_name: &str,
     decay: bool,
@@ -1336,35 +778,35 @@ pub fn generate_generic_c(
     out
 }
 
-pub fn generate_generic_loan_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_loan_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_loan", true)
 }
 
-pub fn generate_generic_loan_mut_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_loan_mut_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_loan_mut", true)
 }
 
-pub fn generate_generic_move_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_move_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_move", true)
 }
 
-pub fn generate_generic_take_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_take_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_take", false)
 }
 
-pub fn generate_generic_take_from_loaned_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_take_from_loaned_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_take_from_loaned", false)
 }
 
-pub fn generate_generic_drop_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_drop_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_drop", false)
 }
 
-pub fn generate_generic_clone_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_clone_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_clone", false)
 }
 
-pub fn generate_take_functions(macro_func: &[FunctionSignature]) -> String {
+fn generate_take_functions(macro_func: &[FunctionSignature]) -> String {
     let mut out = String::new();
     for sig in macro_func {
         let (prefix, _, semantic, _) = split_type_name(&sig.args[0].typename.typename);
@@ -1385,7 +827,7 @@ pub fn generate_take_functions(macro_func: &[FunctionSignature]) -> String {
     out
 }
 
-pub fn generate_move_functions_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_move_functions_c(macro_func: &[FunctionSignature]) -> String {
     let mut out = String::new();
     for sig in macro_func {
         out += &format!(
@@ -1399,7 +841,7 @@ pub fn generate_move_functions_c(macro_func: &[FunctionSignature]) -> String {
     out
 }
 
-pub fn generate_move_functions_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_move_functions_cpp(macro_func: &[FunctionSignature]) -> String {
     let mut out = String::new();
     for sig in macro_func {
         out += &format!(
@@ -1413,19 +855,19 @@ pub fn generate_move_functions_cpp(macro_func: &[FunctionSignature]) -> String {
     out
 }
 
-pub fn generate_generic_null_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_null_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_internal_null", false)
 }
 
-pub fn generate_generic_check_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_check_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_internal_check", true)
 }
 
-pub fn generate_generic_call_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_call_c(macro_func: &[FunctionSignature]) -> String {
     generate_generic_c(macro_func, "z_call", false)
 }
 
-pub fn generate_generic_closure_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_closure_c(macro_func: &[FunctionSignature]) -> String {
     let mut out = "typedef void(*z_closure_drop_callback_t)(void *context);\n".to_string();
     for f in macro_func {
         let callback_typename = f.func_name.clone() + "_callback_t";
@@ -1442,7 +884,7 @@ pub fn generate_generic_closure_c(macro_func: &[FunctionSignature]) -> String {
     out
 }
 
-pub fn generate_generic_recv_c(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_recv_c(macro_func: &[FunctionSignature]) -> String {
     let try_recv_funcs: Vec<FunctionSignature> = macro_func
         .iter()
         .filter(|f| f.func_name.contains("try_recv"))
@@ -1458,7 +900,7 @@ pub fn generate_generic_recv_c(macro_func: &[FunctionSignature]) -> String {
         + generate_generic_c(&recv_funcs, "z_recv", false).as_str()
 }
 
-pub fn generate_generic_cpp(
+fn generate_generic_cpp(
     macro_func: &[FunctionSignature],
     generic_name: &str,
     decay: bool,
@@ -1509,11 +951,11 @@ pub fn generate_generic_cpp(
     out
 }
 
-pub fn generate_generic_loan_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_loan_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_loan", true)
 }
 
-pub fn generate_generic_loan_to_owned_type_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_loan_to_owned_type_cpp(macro_func: &[FunctionSignature]) -> String {
     let mut processed_loaned_types = HashSet::<String>::new();
     let mut out = "template<class T> struct z_loaned_to_owned_type_t {};
 template<class T> struct z_owned_to_loaned_type_t {};"
@@ -1543,43 +985,43 @@ template<> struct z_owned_to_loaned_type_t<{owned}> {{ typedef {loaned} type; }}
     out
 }
 
-pub fn generate_generic_loan_mut_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_loan_mut_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_loan_mut", true)
 }
 
-pub fn generate_generic_drop_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_drop_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_drop", false)
 }
 
-pub fn generate_generic_move_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_move_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_move", true)
 }
 
-pub fn generate_generic_take_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_take_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_take", false)
 }
 
-pub fn generate_generic_take_from_loaned_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_take_from_loaned_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_take_from_loaned", false)
 }
 
-pub fn generate_generic_null_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_null_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_internal_null", false)
 }
 
-pub fn generate_generic_check_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_check_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_internal_check", true)
 }
 
-pub fn generate_generic_call_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_call_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_call", false)
 }
 
-pub fn generate_generic_clone_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_clone_cpp(macro_func: &[FunctionSignature]) -> String {
     generate_generic_cpp(macro_func, "z_clone", false)
 }
 
-pub fn generate_generic_recv_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_recv_cpp(macro_func: &[FunctionSignature]) -> String {
     let try_recv_funcs: Vec<FunctionSignature> = macro_func
         .iter()
         .filter(|f| f.func_name.contains("try_recv"))
@@ -1595,7 +1037,7 @@ pub fn generate_generic_recv_cpp(macro_func: &[FunctionSignature]) -> String {
         + generate_generic_cpp(&recv_funcs, "z_recv", false).as_str()
 }
 
-pub fn generate_generic_closure_cpp(macro_func: &[FunctionSignature]) -> String {
+fn generate_generic_closure_cpp(macro_func: &[FunctionSignature]) -> String {
     // replace function pointer types with using typedefs defined with extern "C" linkage
     let mut out =
         "extern \"C\" using z_closure_drop_callback_t = void(void* context);\n".to_string();
@@ -1618,48 +1060,4 @@ pub fn generate_generic_closure_cpp(macro_func: &[FunctionSignature]) -> String 
 
     out += &generate_generic_cpp(&processed, "z_closure", false);
     out
-}
-
-pub fn evaluate_c_defines_line(line: &str) -> bool {
-    let mut s = line.to_string();
-    for (rust_feature, c_feature) in RUST_TO_C_FEATURES.entries() {
-        s = s.replace(
-            &format!("defined({})", c_feature),
-            match test_feature(rust_feature) {
-                true => "true",
-                false => "false",
-            },
-        );
-    }
-
-    s = s.replace("#if", "");
-    match evalexpr::eval(&s) {
-        Ok(v) => v == evalexpr::Value::from(false),
-        Err(_) => panic!("Failed to evaluate {}", &s),
-    }
-}
-
-pub fn generate_c_headers() {
-    cbindgen::generate(std::env::var("CARGO_MANIFEST_DIR").unwrap())
-        .expect("Unable to generate bindings")
-        .write_to_file(BUGGY_GENERATION_PATH);
-
-    fix_cbindgen(BUGGY_GENERATION_PATH, GENERATION_PATH);
-    std::fs::remove_file(BUGGY_GENERATION_PATH).unwrap();
-
-    preprocess_header(GENERATION_PATH, PREPROCESS_PATH);
-    create_generics_header(PREPROCESS_PATH, "include/zenoh_macros.h");
-    std::fs::remove_file(PREPROCESS_PATH).unwrap();
-
-    configure();
-    let split_guide = SplitGuide::from_yaml(SPLITGUIDE_PATH);
-    split_bindings(&split_guide).unwrap();
-    text_replace(split_guide.rules.iter().map(|(name, _)| name.as_str()));
-
-    fs_extra::copy_items(
-        &["include"],
-        cargo_target_dir(),
-        &fs_extra::dir::CopyOptions::default().overwrite(true),
-    )
-    .expect("include should be copied to CARGO_TARGET_DIR");
 }
