+++ buildrs/utils.rs
@@ -1,30 +1,12 @@
+use phf::phf_set;
 use std::{env, path::PathBuf};
-use phf::phf_map;
 
-pub mod opaque_types;
-pub mod cbindgen;
-
-// See: https://github.com/rust-lang/cargo/issues/9661
-// See: https://github.com/rust-lang/cargo/issues/545
-fn cargo_target_dir() -> PathBuf {
-    let out_dir = PathBuf::from(env::var("OUT_DIR").expect("OUT_DIR should be set"));
-    let profile = env::var("PROFILE").expect("PROFILE should be set");
-
-    let mut target_dir = None;
-    let mut out_dir_path = out_dir.as_path();
-    while let Some(parent) = out_dir_path.parent() {
-        if parent.ends_with(&profile) {
-            target_dir = Some(parent);
-            break;
-        }
-        out_dir_path = parent;
-    }
-
-    target_dir
-        .expect("OUT_DIR should be a child of a PROFILE directory")
-        .to_path_buf()
+pub fn get_build_rs_path() -> PathBuf {
+    let file_path = file!();
+    let mut path_buf = PathBuf::new();
+    path_buf.push(file_path);
+    path_buf.parent().unwrap().parent().unwrap().to_path_buf()
 }
-
 
 pub fn split_type_name(type_name: &str) -> (&str, Option<&str>, &str, &str) {
     let mut split = type_name.split('_');
@@ -45,20 +27,20 @@ pub fn split_type_name(type_name: &str) -> (&str, Option<&str>, &str, &str) {
     (prefix, category, semantic, postfix)
 }
 
-pub static RUST_TO_C_FEATURES: phf::Map<&'static str, &'static str> = phf_map! {
-    "unstable" => "Z_FEATURE_UNSTABLE_API",
-    "shared-memory" => "Z_FEATURE_SHARED_MEMORY",
-    "auth_pubkey" => "Z_FEATURE_AUTH_PUBKEY",
-    "auth_usrpwd" => "Z_FEATURE_AUTH_USRPWD",
-    "transport_multilink" => "Z_FEATURE_TRANSPORT_MULTILINK",
-    "transport_compression" => "Z_FEATURE_TRANSPORT_COMPRESSION",
-    "transport_quic"  => "Z_FEATURE_TRANSPORT_QUIC",
-    "transport_tcp" =>  "Z_FEATURE_TRANSPORT_TCP",
-    "transport_tls" =>  "Z_FEATURE_TRANSPORT_TLS",
-    "transport_udp" =>  "Z_FEATURE_TRANSPORT_UDP",
-    "transport_unixsock-stream" =>  "Z_FEATURE_TRANSPORT_UNIXSOCK_STREAM",
-    "transport_ws" =>  "Z_FEATURE_TRANSPORT_WS",
-    "transport_vsock" => "Z_FEATURE_VSOCK"
+pub static FEATURES: phf::Set<&'static str> = phf_set! {
+    "unstable",
+    "shared-memory",
+    "auth_pubkey",
+    "auth_usrpwd",
+    "transport_multilink",
+    "transport_compression",
+    "transport_quic" ,
+    "transport_tcp",
+    "transport_tls",
+    "transport_udp",
+    "transport_unixsock-stream",
+    "transport_ws",
+    "transport_vsock"
 };
 
 pub fn test_feature(feature: &str) -> bool {
@@ -91,6 +73,29 @@ pub fn test_feature(feature: &str) -> bool {
         "transport_vsock" => true,
         _ => false,
     }
+}
+
+// See: https://github.com/rust-lang/cargo/issues/9661
+// See: https://github.com/rust-lang/cargo/issues/545
+pub fn cargo_target_dir() -> PathBuf {
+    // OUT_DIR is a path to the directory where the build script writes its output.
+    let out_dir = PathBuf::from(env::var("OUT_DIR").expect("OUT_DIR should be set"));
+    // PROFILE is the profile that the build script is being run for. This will be one of "debug" or "release".
+    let profile = env::var("PROFILE").expect("PROFILE should be set");
+
+    let mut target_dir = None;
+    let mut out_dir_path = out_dir.as_path();
+    while let Some(parent) = out_dir_path.parent() {
+        if parent.ends_with(&profile) {
+            target_dir = Some(parent);
+            break;
+        }
+        out_dir_path = parent;
+    }
+
+    target_dir
+        .expect("OUT_DIR should be a child of a PROFILE directory")
+        .to_path_buf()
 }
